server/requirements.txt
flask==2.3.3
flask-jwt-extended==4.5.3
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7 